+++ mosromgr/__init__.py
@@ -2,4 +2,4 @@
 # Copyright 2021 BBC
 # SPDX-License-Identifier: Apache-2.0
 
-__version__ = '0.9.1'
+__version__ = '0.10.0'